refactor `opt_history`
