+++ fedot/core/optimisers/gp_comp/operators/regularization.py
@@ -1,8 +1,9 @@
 from typing import TYPE_CHECKING
 
-from fedot.core.optimisers.gp_comp.individual import Individual, ParentOperator
 from fedot.core.optimisers.gp_comp.operators.operator import PopulationT, EvaluationOperator, Operator
 from fedot.core.optimisers.graph import OptGraph
+from fedot.core.optimisers.opt_history_objects.individual import Individual
+from fedot.core.optimisers.opt_history_objects.parent_operator import ParentOperator
 from fedot.core.optimisers.optimizer import GraphGenerationParams
 from fedot.core.pipelines.node import Node
 from fedot.core.utilities.data_structures import ComparableEnum as Enum
