+++ fedot/core/serializers/serializer.py
@@ -10,6 +10,14 @@ from fedot.core.optimisers.objective.objective import Objective
 MODULE_X_NAME_DELIMITER = '/'
 INSTANCE_OR_CALLABLE = TypeVar('INSTANCE_OR_CALLABLE', object, Callable)
 CLASS_PATH_KEY = '_class_path'
+LEGACY_CLASS_PATHS = {
+    'fedot.core.optimisers.gp_comp.individual/Individual':
+        f'fedot.core.optimisers.opt_history_objects.individual{MODULE_X_NAME_DELIMITER}Individual',
+    'fedot.core.optimisers.gp_comp.individual/ParentOperator':
+        f'fedot.core.optimisers.opt_history_objects.parent_operator{MODULE_X_NAME_DELIMITER}ParentOperator',
+    'fedot.core.optimisers.opt_history/OptHistory':
+        f'fedot.core.optimisers.opt_history_objects.opt_history{MODULE_X_NAME_DELIMITER}OptHistory',
+}
 
 
 class Serializer(JSONEncoder, JSONDecoder):
@@ -31,10 +39,10 @@ class Serializer(JSONEncoder, JSONDecoder):
             from fedot.core.dag.graph import Graph
             from fedot.core.dag.graph_node import GraphNode
             from fedot.core.operations.operation import Operation
-            from fedot.core.optimisers.gp_comp.individual import Individual
+            from fedot.core.optimisers.opt_history_objects.individual import Individual
+            from fedot.core.optimisers.opt_history_objects.opt_history import OptHistory
+            from fedot.core.optimisers.opt_history_objects.parent_operator import ParentOperator
             from fedot.core.optimisers.graph import OptGraph, OptNode
-            from fedot.core.optimisers.opt_history import OptHistory
-            from fedot.core.optimisers.gp_comp.individual import ParentOperator
             from fedot.core.utilities.data_structures import ComparableEnum
 
             from .coders import (
