+++ fedot/utilities/project_import_export.py
@@ -5,7 +5,7 @@ from typing import Optional, Tuple, Union
 
 from fedot.core.data.data import InputData
 from fedot.core.log import LoggerAdapter, default_log
-from fedot.core.optimisers.opt_history import OptHistory
+from fedot.core.optimisers.opt_history_objects.opt_history import OptHistory
 from fedot.core.pipelines.pipeline import Pipeline
 from fedot.core.utils import default_fedot_data_dir
