+++ test/integration/api_params/test_main_api_params.py
@@ -5,7 +5,7 @@ from typing import Callable, Union
 import pytest
 
 from fedot.api.main import Fedot
-from fedot.core.optimisers.opt_history import OptHistory
+from fedot.core.optimisers.opt_history_objects.opt_history import OptHistory
 from fedot.core.repository.tasks import TsForecastingParams
 from test.unit.api.test_main_api import get_dataset
