+++ test/unit/optimizer/test_elitism_operator.py
@@ -1,15 +1,13 @@
-from typing import TYPE_CHECKING
-
 import pytest
 
 from fedot.core.optimisers.adapters import PipelineAdapter
 from fedot.core.optimisers.gp_comp.evaluation import SimpleDispatcher
-from fedot.core.optimisers.gp_comp.individual import Individual
+from fedot.core.optimisers.gp_comp.gp_params import GPGraphOptimizerParameters
 from fedot.core.optimisers.gp_comp.operators.elitism import Elitism, ElitismTypesEnum
+from fedot.core.optimisers.opt_history_objects.individual import Individual
 from test.unit.optimizer.test_evaluation import prepared_objective
 from test.unit.pipelines.test_node_cache import pipeline_first, pipeline_second, pipeline_third, pipeline_fourth, \
     pipeline_fifth
-from fedot.core.optimisers.gp_comp.gp_params import GPGraphOptimizerParameters
 
 
 @pytest.fixture()
