+++ test/unit/optimizer/test_evaluation.py
@@ -5,8 +5,8 @@ import pytest
 from fedot.core.optimisers.adapters import PipelineAdapter
 from fedot.core.optimisers.fitness import Fitness, null_fitness
 from fedot.core.optimisers.gp_comp.evaluation import MultiprocessingDispatcher, SimpleDispatcher
-from fedot.core.optimisers.gp_comp.individual import Individual
 from fedot.core.optimisers.objective import Objective
+from fedot.core.optimisers.opt_history_objects.individual import Individual
 from fedot.core.optimisers.timer import OptimisationTimer
 from fedot.core.pipelines.pipeline import Pipeline
 from fedot.core.repository.quality_metrics_repository import ClassificationMetricsEnum
